% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_plot.R
\name{acm_plot}
\alias{acm_plot}
\title{Title}
\usage{
acm_plot(df, dims, name_ind = 0, qtsup = NULL)
}
\arguments{
\item{df}{a dataframe with qualitatives features}

\item{dims}{dimensions to print}

\item{name_ind}{rownames to apply}
}
\value{
corrplot with the contributions of features for each dimensions and a biplot to vizualize acm
}
\description{
Title
}
\examples{

}
