% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test_all.R
\name{chisq_test_all}
\alias{chisq_test_all}
\title{Title}
\usage{
chisq_test_all(name_var_group, data)
}
\arguments{
\item{name_var_group}{string with represent the name of target column}

\item{data}{dataframe with qualitative columns (use extract_ql(dataframe) to extract qualitative variables)}
}
\value{
a dataframe with a summary of chisq.test between the target variable and others qualitatives variables
}
\description{
Title
}
\examples{
#(change example)
#data=read.csv2("Données-R-1.csv", sep=";", row.names = 1)
#Data.qual<- extract_ql(data)
#chisq_test_all("Continent",Data.qual)
}
