% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R, R/classe_multivariate.R,
%   R/data_type.R
\name{data_type}
\alias{data_type}
\title{Title}
\usage{
data_type(X)

data_type(X)

data_type(X)
}
\arguments{
\item{X}{dataframe or matrix}
}
\value{


the type of X: there are 5 possibilities: quantitative, quantitatives,
qualitative, qualitatives, quantitative-qualitative

the type of X: there are 5 possibilities: quantitative, quantitatives,
qualitative, qualitatives, quantitative-qualitative
}
\description{
Title

Title

Title
}
