% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_prof.R
\name{desc_profils}
\alias{desc_profils}
\title{desc_prof}
\usage{
desc_profils(data, name_var_group, name_var_comp)
}
\arguments{
\item{name_var_group}{the target variable (group variable)}

\item{name_var_comp}{the other qualitative variable}
}
\value{
contingence table, and profils
}
\description{
desc_prof
}
