% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R, R/size_effect.R
\name{desc_size_effect}
\alias{desc_size_effect}
\title{Title}
\usage{
desc_size_effect(df, target_grp, var_expli)

desc_size_effect(df, target_grp, var_expli)
}
\arguments{
\item{df}{the dataframe}

\item{target_grp}{}

\item{object}{your Univariate object}

\item{ind_var_exp}{the qualitative explonatory feature}
}
\value{
the vtest value for each class(group feature)/modality(explonatory feature)

a description (distribution) of size effect between a group choose by the user and an other variable(for each modality)
}
\description{
Title

Title
}
\examples{

#data = data_0[c(14:16,18)]
#desc_size_effect(data,'liberte_internet','Régime_simp')


}
