% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport_correlation.R
\name{eta2}
\alias{eta2}
\title{Title}
\usage{
eta2(x, g)
}
\arguments{
\item{x}{a variable with quantitative value}

\item{g}{a factor such that length(x)=length(g)}
}
\value{
rapport correlation
}
\description{
Title
}
