% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_test_all.R
\name{fisher_test_all}
\alias{fisher_test_all}
\title{Title}
\usage{
fisher_test_all(data, g)
}
\arguments{
\item{data}{a dataframe or matrix}

\item{g}{cluster identity}
}
\value{
a dataframe with a summary of Fisher test between the target variable and others quanlitatives variables
}
\description{
Title
}
