% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R, R/kmean_rand_index.R
\name{kmean_rand_index}
\alias{kmean_rand_index}
\title{Title}
\usage{
kmean_rand_index(X, y)

kmean_rand_index(X, y)
}
\arguments{
\item{X}{dataframe or matrix}

\item{y}{cluster identity by using another algorithm or anothers explanatory variables}

\item{object}{your Univariate object}
}
\value{


rand index between the result of kmean and y
}
\description{
Title

Title
}
