% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{kmean_rand_index_univ}
\alias{kmean_rand_index_univ}
\title{Title}
\usage{
kmean_rand_index_univ(object)
}
\arguments{
\item{object}{your Univariate object}
}
\value{
rand index between the result of kmean and y
}
\description{
Title
}
\examples{
kmean_rand_index_univ(Univariate_object(esoph,1))
}
