% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{m_acm_plot_multi}
\alias{m_acm_plot_multi}
\title{Plot ACM multivariate}
\usage{
m_acm_plot_multi(object, dims = c(1, 2), name_ind = 0, qtsup = NULL)
}
\arguments{
\item{object}{your Multivariate object}

\item{dims}{dimensions to print}

\item{name_ind}{rownames to apply}

\item{qtsup}{if users want to add a quantitative features to acm (could be a number or vector of numbers)}
}
\value{
corrplot with the contributions of features for each dimensions and a biplot to vizualize acm
}
\description{
Plot ACM multivariate
}
\examples{
m_acm_plot_multi(multivariate_object(CO2[,1:3],1))
}
