% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R
\name{m_acp_2_axes}
\alias{m_acp_2_axes}
\title{Values ACP with 2 dimensions}
\usage{
m_acp_2_axes(X, i = 1, j = 2, rescale = FALSE)
}
\arguments{
\item{rescale}{}
}
\value{
the values of ACP with 2 dimensions
}
\description{
Values ACP with 2 dimensions
}
\examples{
m_acp_2_axes(iris[,-5])
}
