% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{m_data_type}
\alias{m_data_type}
\title{All type possible in the dataframe}
\usage{
m_data_type(X)
}
\arguments{
\item{X}{dataframe or matrix}
}
\value{
the type of X: there are 5 possibilities: quantitative, quantitatives,
qualitative, qualitatives, quantitative-qualitative
}
\description{
All type possible in the dataframe
}
\examples{
#m_data_type(iris)
}
