% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R
\name{m_DB_index}
\alias{m_DB_index}
\title{Davies-Boulin Indice}
\usage{
m_DB_index(object, method = "encoding", rescale = FALSE)
}
\arguments{
\item{object}{your Multivariate object}
}
\value{
Indice de Davies-Bouldin
}
\description{
Davies-Boulin Indice
}
\examples{
m_DB_index(multivariate_object(infert,1))
}
