% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{m_dummy_data}
\alias{m_dummy_data}
\title{Title}
\usage{
m_dummy_data(data, rescale = FALSE)
}
\arguments{
\item{data}{a datafram}

\item{rescale}{a boolean}
}
\value{
data encoding with rescale or not
}
\description{
Title
}
