% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{m_kmean_clustering_plot}
\alias{m_kmean_clustering_plot}
\title{Title}
\usage{
m_kmean_clustering_plot(object, i = 1, j = 2, interact = TRUE)
}
\arguments{
\item{object}{your Multivariate object}

\item{j}{}
}
\value{
a plot about the cluster on the first two dimensions
}
\description{
Title
}
\examples{
m_kmean_clustering_plot(multivariate_object(infert,1))

}
