% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R
\name{m_kmean_rand_ajusted}
\alias{m_kmean_rand_ajusted}
\title{Compare two partitions, real classes vs kmeans classes (with ajusted indice)}
\usage{
m_kmean_rand_ajusted(object, rescale = FALSE)
}
\arguments{
\item{object}{your Multivariate object}
}
\value{
ajusted rand index between the result of kmean and y
}
\description{
Compare two partitions, real classes vs kmeans classes (with ajusted indice)
}
\examples{
m_kmean_rand_ajusted(multivariate_object(infert,1))

}
