% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{m_matrix_distance}
\alias{m_matrix_distance}
\title{Distance Matrix}
\usage{
m_matrix_distance(X, d)
}
\arguments{
\item{X}{a matrix or dataframe with explanatory variables}

\item{d}{the distance measure to be used and this must be "euclidean" or "L1"}
}
\value{
the distance matrix computed by using the euclidean distance or L1 distance between the rows of X.
}
\description{
Distance Matrix
}
\examples{
#m_matrix_distance(iris[,c("Sepal.Length","Sepal.Width")],d="euclidean")
}
