% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R
\name{m_mean_distance}
\alias{m_mean_distance}
\title{Mean distance}
\usage{
m_mean_distance(X, y)
}
\arguments{
\item{X}{a numeric symmetric matrix containing the pairwise distance between the rows of a data frame}

\item{y}{a factor such that length(X)=length(y)}
}
\value{
a data frame containing the average distance of each row to all observations of each cluster C
}
\description{
Mean distance
}
\examples{
 m_mean_distance(m_matrix_distance(iris[,c("Sepal.Length","Sepal.Width")],d="euclidean"), iris[,'Species'])

}
