% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{m_R2_multivariate}
\alias{m_R2_multivariate}
\title{Rapport Correlation Multivariate}
\usage{
m_R2_multivariate(object, method = "encoding", rescale = FALSE)
}
\arguments{
\item{object}{your Multivariate object}

\item{method}{method used}
}
\value{
rapport correlation
}
\description{
Rapport Correlation Multivariate
}
\examples{
m_R2_multivariate(multivariate_object(infert,1))
}
