% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R, R/classe_multivariate.R
\name{m_R2_multivariate_multi}
\alias{m_R2_multivariate_multi}
\title{Rapport Correlation Multivariate}
\usage{
m_R2_multivariate_multi(object, method = "encoding")

m_R2_multivariate_multi(object, method = "encoding")
}
\arguments{
\item{object}{your Multivariate object}

\item{method}{method used}
}
\value{
rapport correlation

rapport correlation
}
\description{
Rapport Correlation Multivariate

Rapport Correlation Multivariate
}
\examples{
m_R2_multivariate_multi(multivariate_object(infert,1))
m_R2_multivariate_multi(multivariate_object(infert,1))
}
