% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R
\name{m_rand_ajusted}
\alias{m_rand_ajusted}
\title{Calculate adjusted rand index}
\usage{
m_rand_ajusted(g1, g2)
}
\arguments{
\item{g1}{first cluster}

\item{g2}{second cluster}
}
\value{
ajusted Rand index measure to compare the similarity of two clustering.
}
\description{
Calculate adjusted rand index
}
\examples{

}
