% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R
\name{m_rand_index}
\alias{m_rand_index}
\title{Calculate rand index}
\usage{
m_rand_index(g1, g2)
}
\arguments{
\item{g1}{first cluster}

\item{g2}{second cluster}
}
\value{
Rand index measure to compare the similarity of two clustering.
}
\description{
Calculate rand index
}
\examples{

}
