% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R
\name{m_sil_pca_plot_multi}
\alias{m_sil_pca_plot_multi}
\title{Title}
\usage{
m_sil_pca_plot_multi(object, i = 1, j = 2, rescale = FALSE, d = "euclidean")
}
\arguments{
\item{object}{your Multivariate object}

\item{d}{method used}
}
\value{

}
\description{
Title
}
\examples{
m_sil_pca_plot_multi(multivariate_object(iris,5))

}
