% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R
\name{m_silhouette_ind}
\alias{m_silhouette_ind}
\title{Calculate index silhouette}
\usage{
m_silhouette_ind(object, rescale = FALSE, d = "euclidean")
}
\arguments{
\item{object}{your Multavariate object}

\item{d}{method used}
}
\value{
Silhouette Coefficient of each row
}
\description{
Calculate index silhouette
}
\examples{
m_silhouette_ind(multivariate_object(iris,5))

}
