% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{m_silhouette_plot}
\alias{m_silhouette_plot}
\title{Title}
\usage{
m_silhouette_plot(object, rescale = FALSE, d = "euclidean", interact = TRUE)
}
\arguments{
\item{object}{your Multivariate object}

\item{d}{method used}
}
\value{
the silhouette plot
}
\description{
Title
}
\examples{
m_silhouette_plot(multivariate_object(iris,5))

}
