% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{m_test.value}
\alias{m_test.value}
\title{Title}
\usage{
m_test.value(object, i = 1)
}
\arguments{
\item{object}{your Multivariate object}

\item{i}{i-th cluster}
}
\value{
a dataframe with a summary of value test between the explanatory variables and and the i-th cluster
}
\description{
Title
}
\examples{
m_test.value(multivariate_object(infert,1))
}
