% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R
\name{m_test_value}
\alias{m_test_value}
\title{Title}
\usage{
m_test_value(object, i = 1)
}
\arguments{
\item{object}{your Multivariate object}

\item{i}{i-th cluster}
}
\value{
a dataframe with a summary of value test between the explanatory variables and and the i-th cluster
}
\description{
Title
}
\examples{
m_test_value(multivariate_object(infert,1))
}
