% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{m_type_variable}
\alias{m_type_variable}
\title{Verify the type of variables}
\usage{
m_type_variable(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the type of x
}
\description{
Verify the type of variables
}
\examples{
#m_type_variable(iris$Sepal.Length)
}
