% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R, R/classe_multivariate.R,
%   R/mean_distance.R
\name{mean_distance}
\alias{mean_distance}
\title{Title}
\usage{
mean_distance(X, y)

mean_distance(X, y)

mean_distance(X, y)
}
\arguments{
\item{X}{a numeric symmetric matrix containing the pairwise distance between the rows of a data frame}

\item{y}{a factor such that length(X)=length(y)}
}
\value{


a data frame containing the average distance of each row to all observations of each cluster C

a data frame containing the average distance of each row to all observations of each cluster C
}
\description{
Title

Title

Title
}
