% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_var.R
\name{mode_var}
\alias{mode_var}
\title{Title}
\usage{
mode_var(x)
}
\arguments{
\item{x}{a column with qualitative values or discrete values}
}
\value{
the value that appears the most (mode)
}
\description{
Title
}
