% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{multivariate_object}
\alias{multivariate_object}
\title{Title}
\usage{
multivariate_object(df, ind_group_class)
}
\arguments{
\item{df}{your dataframe including explanatory features and the group feature(target)}

\item{ind_group_class}{the indice of your group feature}
}
\value{
an object which is going to be use to analyse clustering
}
\description{
Title
}
