% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport_corr_multivariate_quanti.R
\name{R2_multivariate}
\alias{R2_multivariate}
\title{Title}
\usage{
R2_multivariate(data, g, method = "encoding")
}
\arguments{
\item{data}{a datafram or matrix with quantitative value}

\item{g}{a factor such that length(x)=nrow(data)}
}
\value{
rapport correlation
}
\description{
Title
}
