% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R, R/rand_index.R
\name{rand_index}
\alias{rand_index}
\title{Title}
\usage{
rand_index(g1, g2)

rand_index(g1, g2)
}
\arguments{
\item{g1}{first cluster}

\item{g2}{second cluster}
}
\value{


Rand index measure to compare the similarity of two clustering.
}
\description{
Title

Title
}
