% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalisation_quanti.R, R/shapiro_test.R
\name{shapiro_test}
\alias{shapiro_test}
\title{Title}
\usage{
shapiro_test(data)

shapiro_test(data)
}
\arguments{
\item{data}{a dataframe}

\item{pop_b}{}
}
\value{


a dataframe with a summary of shapiro.test
}
\description{
Title

Title
}
\examples{
#data=read.csv2("Données-R-1.csv", sep=";", row.names = 1)
#Data.quanti<- detection_quanti(data)
#shapiro_test(Data.quanti)

}
