% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R, R/silhouette.R
\name{silhouette_ind}
\alias{silhouette_ind}
\title{Title}
\usage{
silhouette_ind(X, y, rescale = FALSE, d = "euclidean")

silhouette_ind(X, y, rescale = FALSE, d = "euclidean")
}
\arguments{
\item{X}{a matrix or dataframe with explanatory variables}

\item{y}{predicted labels for each row of X. It should be a factor}

\item{d}{the distance measure to be used and this must be "euclidean" or "L1"}

\item{object}{your Univariate object}
}
\value{


Silhouette Coefficient of each row
}
\description{
Title

Title
}
\examples{

}
