% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{silhouette_ind_multi}
\alias{silhouette_ind_multi}
\title{Title}
\usage{
silhouette_ind_multi(object, rescale = FALSE, d = "euclidean")
}
\arguments{
\item{d}{}
}
\value{
Silhouette Coefficient of each row
}
\description{
Title
}
\examples{

}
