% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette_plot.R
\name{silhouette_plot}
\alias{silhouette_plot}
\title{Title}
\usage{
silhouette_plot(X, y, rescale = FALSE, d = "euclidean")
}
\arguments{
\item{X}{a matrix or dataframe with explanatory variables}

\item{y}{predicted labels for each row of X. It should be a factor}

\item{d}{the distance measure to be used and this must be "euclidean" or "L1"}
}
\value{
the silhouette plot
}
\description{
Title
}
