% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{silhouette_plot_univ}
\alias{silhouette_plot_univ}
\title{Title}
\usage{
silhouette_plot_univ(object, rescale = FALSE, d = "euclidean")
}
\arguments{
\item{object}{your Univariate object}

\item{d}{the distance measure to be used and this must be "euclidean" or "L1"}

\item{y}{predicted labels for each row of X. It should be a factor}
}
\value{
the silhouette plot
}
\description{
Title
}
\examples{
#silhouette_plot_univ(Univariate_object(esoph,1))
}
