% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_value_quanti.R
\name{test.value}
\alias{test.value}
\title{Title}
\usage{
test.value(data, g, i = 1)
}
\arguments{
\item{data}{a dataframe or matrix with quantitatives values}

\item{g}{cluster identity}

\item{i}{i-th cluster}
}
\value{
a dataframe with a summary of value test between the explanatory variables and and the i-th cluster
}
\description{
Title
}
