% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R, R/generalisation_quanti.R
\name{ttest_all}
\alias{ttest_all}
\title{Title}
\usage{
ttest_all(df, var_groupe)

ttest_all(df, var_groupe)
}
\arguments{
\item{var_groupe}{}

\item{object}{your Univariate object}
}
\value{



}
\description{
Title

Title
}
