% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R, R/type_variable.R
\name{type_variable}
\alias{type_variable}
\title{Title}
\usage{
type_variable(x)

type_variable(x)
}
\arguments{
\item{x}{a vector}
}
\value{


the type of x
}
\description{
Title

Verify the type of variables
}
