% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_acp_2_axes}
\alias{u_acp_2_axes}
\title{Title}
\usage{
u_acp_2_axes(X, i = 1, j = 2, rescale = FALSE)
}
\arguments{
\item{X}{a dataframe}

\item{i}{first dim}

\item{j}{second dim}

\item{rescale}{a boolean}
}
\value{
dataframe with ACP results for 2 axes
}
\description{
Title
}
