% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_afc_plot}
\alias{u_afc_plot}
\title{Vizualise modality and class on the same plan}
\usage{
u_afc_plot(object, ind_var_exp)
}
\arguments{
\item{object}{your Univariate object}

\item{ind_var_exp}{the indice of the explanatory variable (only qualitative)}
}
\value{
a factorial projection between group variable and explanatory variable
}
\description{
Vizualise modality and class on the same plan
}
\examples{
u_afc_plot((Univariate_object(esoph,1)),3)
}
