% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_chisq_test_all}
\alias{u_chisq_test_all}
\title{Title}
\usage{
u_chisq_test_all(object)
}
\arguments{
\item{object}{your Univariate object}
}
\value{
all the chisq.test value between your qualitatives variables and your group variable (with interpretation)
}
\description{
Title
}
\examples{
u_chisq_test_all(Univariate_object(esoph,1))
}
