% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_desc_profils}
\alias{u_desc_profils}
\title{See raws and columns profils, and a barplot}
\usage{
u_desc_profils(object, ind_var_exp, interact = TRUE)
}
\arguments{
\item{object}{your Univariate object}

\item{ind_var_exp}{the indice of the explanatory variable (only qualitative)}

\item{interact}{if TRUE then an interactive vizualisation is generate with ggplotly, else there is just a classic plot}
}
\value{
contingency table between group and explanatory variables and the rows and columns profils
}
\description{
See raws and columns profils, and a barplot
}
\examples{
u_desc_profils((Univariate_object(esoph,1)),3)
}
