% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_dummy_data}
\alias{u_dummy_data}
\title{Title}
\usage{
u_dummy_data(data, rescale = FALSE)
}
\arguments{
\item{data}{a datafram}

\item{rescale}{a boolean}
}
\value{
data encoding with rescale or not
}
\description{
Title
}
