% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_explonarory_test}
\alias{u_explonarory_test}
\title{Title}
\usage{
u_explonarory_test(object, i = 1)
}
\arguments{
\item{object}{your Univariate object}

\item{i}{the number of your target group}
}
\value{
the explanatory variable which caracterize the most the class of your group variable you choose
}
\description{
Title
}
\examples{
u_explonarory_test(Univariate_object(esoph,1))
}
