% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_fisher_test_all}
\alias{u_fisher_test_all}
\title{Title}
\usage{
u_fisher_test_all(object)
}
\arguments{
\item{object}{your Univariate object}
}
\value{
a dataframe with a summary of Fisher test between the target variable and others quanlitatives variables
}
\description{
Title
}
\examples{
u_fisher_test_all(Univariate_object(iris,5))
}
