% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_kmean_rand_index}
\alias{u_kmean_rand_index}
\title{Title}
\usage{
u_kmean_rand_index(object)
}
\arguments{
\item{object}{your Univariate object}
}
\value{
rand index between the result of kmean and your group variable
}
\description{
Title
}
\examples{
u_kmean_rand_index(Univariate_object(iris,5))
}
