% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_matrix_distance}
\alias{u_matrix_distance}
\title{Title}
\usage{
u_matrix_distance(X, d)
}
\arguments{
\item{X}{a matrix or dataframe with explanatory variables}

\item{d}{the distance measure to be used and this must be "euclidean" or "L1"}
}
\value{
the distance matrix computed by using the euclidean distance or L1 distance between the rows of X.
}
\description{
Title
}
