% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_Mosaic_plot}
\alias{u_Mosaic_plot}
\title{Create mosaic plot}
\usage{
u_Mosaic_plot(explanatory, cluster)
}
\arguments{
\item{explanatory}{your explanatory variable}

\item{cluster}{}
}
\value{

}
\description{
Create mosaic plot
}
