% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_plot_size_effect}
\alias{u_plot_size_effect}
\title{Plot the size effect with a mosaic plot}
\usage{
u_plot_size_effect(object, name_var_exp, interact = TRUE)
}
\arguments{
\item{object}{your univariate object}

\item{interact}{if TRUE then an interactive vizualisation is generate with ggplotly, else there is just a classic plot}

\item{ind_var_exp}{the indice of your explanatory variable(only qualitative)}
}
\value{
a mosaic plot which is the distribution of your explanatory variable by class
}
\description{
Plot the size effect with a mosaic plot
}
\examples{
u_plot_size_effect((Univariate_object(esoph,1)),2)
}
