% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_rand_index}
\alias{u_rand_index}
\title{Title}
\usage{
u_rand_index(g1, g2)
}
\arguments{
\item{g1}{first cluster}

\item{g2}{second cluster}
}
\value{
Rand index measure to compare the similarity of two clustering.
}
\description{
Title
}
