% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_silhouette_ind}
\alias{u_silhouette_ind}
\title{Title}
\usage{
u_silhouette_ind(object, rescale = FALSE, d = "euclidean")
}
\arguments{
\item{object}{your Univariate object}

\item{rescale}{rescale or not}

\item{d}{distance type}
}
\value{
Silhouette Coefficient of each row
}
\description{
Title
}
\examples{
u_silhouette_ind(Univariate_object(iris,5))
}
