% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_silhouette_plot}
\alias{u_silhouette_plot}
\title{Title}
\usage{
u_silhouette_plot(object, rescale = FALSE, d = "euclidean", interact = TRUE)
}
\arguments{
\item{object}{your Univariate object}

\item{rescale}{rescale or not}

\item{d}{the distance measure to be used and this must be "euclidean" or "L1"}

\item{interact}{if TRUE then an interactive vizualisation is generate with ggplotly, else there is just a classic plot}
}
\value{
the silhouette plot
}
\description{
Title
}
\examples{
u_silhouette_plot(Univariate_object(iris,5))
}
