% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_test_value}
\alias{u_test_value}
\title{Title}
\usage{
u_test_value(object, i = 1)
}
\arguments{
\item{object}{your Univariate object}

\item{i}{the number of your target group}
}
\value{
the explanatory variable which caracterize the most the class of your group variable you choose
}
\description{
Title
}
\examples{
u_test_value(Univariate_object(iris[,c(1,5)],2))
}
