% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_corr.R
\name{viz_corr}
\alias{viz_corr}
\title{Title}
\usage{
viz_corr(df)
}
\arguments{
\item{df}{a dataframe with numeric variables}
}
\value{
the correlation matrix using ggcorrplot vizualisation
}
\description{
Title
}
\examples{
viz_corr(mtcars)
}
