% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R, R/afc_plot.R
\name{afc_plot}
\alias{afc_plot}
\title{Title}
\usage{
afc_plot(contingence)

afc_plot(contingence)
}
\arguments{
\item{contingence}{: a contingence table}

\item{object}{your Univariate object}

\item{ind_var_exp}{}
}
\value{



}
\description{
Title

Title
}
\examples{

}
