% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R, R/desc_prof.R
\name{desc_profils}
\alias{desc_profils}
\title{Title}
\usage{
desc_profils(data, name_var_group, name_var_comp)

desc_profils(data, name_var_group, name_var_comp)
}
\arguments{
\item{name_var_group}{the target variable (group variable)}

\item{name_var_comp}{the other qualitative variable}

\item{object}{your Univariate object}

\item{ind_var_exp}{}
}
\value{


contingence table, and profils
}
\description{
Title

desc_prof
}
\examples{
#desc_profils(data,"heureux_quali","Region")
}
