% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_effect.R
\name{desc_size_effect}
\alias{desc_size_effect}
\title{Title}
\usage{
desc_size_effect(df, var_expli, target_grp)
}
\arguments{
\item{df}{the dataframe}

\item{target_grp}{}
}
\value{
a description (distribution) of size effect between a group choose by the user and an other variable(for each modality)
}
\description{
Title
}
