% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{desc_size_effect}
\alias{desc_size_effect}
\title{Title}
\usage{
desc_size_effect(object, ind_var_exp)
}
\arguments{
\item{object}{your Univariate object}

\item{ind_var_exp}{the qualitative explonatory variable}
}
\value{
the vtest value for each class(group variable)/modality(explonatory variable)
}
\description{
Title
}
\examples{
desc_size_effect((Univariate_object(esoph,1)),3)

}
