% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_data.R
\name{dummy_data}
\alias{dummy_data}
\title{Title}
\usage{
dummy_data(data, rescale = FALSE)
}
\arguments{
\item{data}{a datafram}

\item{rescale}{a boolean}
}
\value{
data encoding with rescale or not
}
\description{
Title
}
