% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{m_DB_index}
\alias{m_DB_index}
\title{Indice de Davies-Boulin}
\usage{
m_DB_index(object, method = "encoding")
}
\arguments{
\item{object}{your Multivariate object}
}
\value{
Indice de Davies-Bouldin
}
\description{
Indice de Davies-Boulin
}
\examples{
m_DB_index(multivariate_object(infert,1))
}
