% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R
\name{m_kmean_clustering_plot_multi}
\alias{m_kmean_clustering_plot_multi}
\title{Title}
\usage{
m_kmean_clustering_plot_multi(object, i = 1, j = 2)
}
\arguments{
\item{object}{your Multivariate object}

\item{j}{}
}
\value{
a plot about the cluster on the first two dimensions
}
\description{
Title
}
\examples{
m_kmean_clustering_plot_multi(multivariate_object(infert,1))

}
