% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{m_kmean_rand_ajusted}
\alias{m_kmean_rand_ajusted}
\title{Title}
\usage{
m_kmean_rand_ajusted(object, rescale = FALSE)
}
\arguments{
\item{object}{your Multivariate object}
}
\value{
ajusted rand index between the result of kmean and y
}
\description{
Title
}
\examples{
m_kmean_rand_ajusted(multivariate_object(infert,1))

}
