% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R
\name{m_kmean_rand_index_multi}
\alias{m_kmean_rand_index_multi}
\title{Title}
\usage{
m_kmean_rand_index_multi(object)
}
\arguments{
\item{object}{your Multivariate object}
}
\value{
rand index between the result of kmean and y
}
\description{
Title
}
\examples{
m_kmean_rand_index_multi(multivariate_object(infert,1))

}
