% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R, R/classe_multivariate.R
\name{m_mean_distance}
\alias{m_mean_distance}
\title{Title}
\usage{
m_mean_distance(X, y)

m_mean_distance(X, y)
}
\arguments{
\item{X}{a numeric symmetric matrix containing the pairwise distance between the rows of a data frame}

\item{y}{a factor such that length(X)=length(y)}
}
\value{
a data frame containing the average distance of each row to all observations of each cluster C

a data frame containing the average distance of each row to all observations of each cluster C
}
\description{
Title

Title
}
\examples{
 m_mean_distance(m_matrix_distance(iris[,c("Sepal.Length","Sepal.Width")],d="euclidean"), iris[,'Species'])

 m_mean_distance(m_matrix_distance(iris[,c("Sepal.Length","Sepal.Width")],d="euclidean"), iris[,'Species'])

}
