% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R
\name{m_sil_pca_plot}
\alias{m_sil_pca_plot}
\title{Plot silhouette index with Component analysis}
\usage{
m_sil_pca_plot(
  object,
  i = 1,
  j = 2,
  rescale = FALSE,
  d = "euclidean",
  interact = TRUE
)
}
\arguments{
\item{object}{your Multivariate object}

\item{d}{method used}
}
\value{

}
\description{
Plot silhouette index with Component analysis
}
\examples{
m_sil_pca_plot(multivariate_object(iris,5))

}
