% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate_Class.R
\name{m_silhouette_plot_multi}
\alias{m_silhouette_plot_multi}
\title{Title}
\usage{
m_silhouette_plot_multi(object, rescale = FALSE, d = "euclidean")
}
\arguments{
\item{object}{your Multivariate object}

\item{d}{method used}
}
\value{
the silhouette plot
}
\description{
Title
}
\examples{
m_silhouette_plot_multi(multivariate_object(iris,5))

}
