% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_multivariate.R, R/type_variable.R
\name{type_variable}
\alias{type_variable}
\title{Verify the type of variables}
\usage{
type_variable(x)

type_variable(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the type of x

the type of x
}
\description{
Verify the type of variables

Verify the type of variables
}
