% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_afc_plot}
\alias{u_afc_plot}
\title{Title}
\usage{
u_afc_plot(object, ind_var_exp)
}
\arguments{
\item{object}{your Univariate object}

\item{ind_var_exp}{the indice of the explonatory variable (only qualitative)}
}
\value{
a factorial projection between group variable and explonatory variable
}
\description{
Title
}
\examples{
u_afc_plot((Univariate_object(esoph,1)),3)
}
