% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_desc_profils}
\alias{u_desc_profils}
\title{Title}
\usage{
u_desc_profils(object, ind_var_exp)
}
\arguments{
\item{object}{your Univariate object}

\item{ind_var_exp}{the indice of the explonatory variable (only qualitative)}
}
\value{
contingency table between group and explonatory variables and the rows and columns profils
}
\description{
Title
}
\examples{
u_desc_profils((Univariate_object(esoph,1)),3)
}
