% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_desc_size_effect}
\alias{u_desc_size_effect}
\title{Function to see if some class are over-representend or not}
\usage{
u_desc_size_effect(object, ind_var_exp)
}
\arguments{
\item{object}{your Univariate object}

\item{ind_var_exp}{the qualitative explanatory variable}
}
\value{
the vtest value for each class(group variable)/modality(explanatory variable)
}
\description{
Function to see if some class are over-representend or not
}
\examples{
u_desc_size_effect((Univariate_object(esoph,1)),3)

}
