% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_plot_size_effect}
\alias{u_plot_size_effect}
\title{Title}
\usage{
u_plot_size_effect(object, name_var_exp)
}
\arguments{
\item{object}{your univariate object}

\item{ind_var_exp}{the indice of your explanatory variable(only qualitative)}
}
\value{
a mosaic plot which is the distribution of your explanatory variable by class
}
\description{
Title
}
\examples{
u_plot_size_effect((Univariate_object(esoph,1)),2)
}
