% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_shapiro_test}
\alias{u_shapiro_test}
\title{Shapiro test}
\usage{
u_shapiro_test(pop_a, pop_b)
}
\arguments{
\item{pop_a}{the first population to test}

\item{pop_b}{the second population to test}
}
\value{
TRUE if gaussian hypothesis is verified or FALSE if it's not verified
}
\description{
Shapiro test
}
