% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{u_sil_pca_plot}
\alias{u_sil_pca_plot}
\title{Title}
\usage{
u_sil_pca_plot(object, i = 1, j = 2, rescale = FALSE, d = "euclidean")
}
\arguments{
\item{d}{}
}
\value{

}
\description{
Title
}
\examples{
u_sil_pca_plot(Univariate_object(iris,5))
}
