% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_Class.R
\name{Univariate_object}
\alias{Univariate_object}
\title{Title}
\usage{
Univariate_object(df, ind_group_class)
}
\arguments{
\item{df}{your dataframe including explanatory variables and the group variable(target)}

\item{ind_group_class}{the indice of your group variable}
}
\value{
an object which is going to be use to analyze clustering
}
\description{
Title
}
\examples{
Univariate_object(esoph,1)

}
